refactor: move variant to its own dir
