+++ server/src/application/sentences/repository/variant-repository.ts
@@ -1,6 +1,6 @@
 import { option as O, taskEither as TE } from 'fp-ts'
 import { ApplicationError } from '../../types/error'
-import { Variant } from '../../../core/types/variant'
+import { Variant } from '../../../core/variants/variant'
 import { queryDb } from '../../../infrastructure/db/mysql'
 import { pipe } from 'fp-ts/lib/function'
 import { createDatabaseError } from '../../helper/error-helper'
