+++ server/src/core/variants/variant.ts
